Create pg_backup.sh
Creates a backup of cluster globals and every database in the cluster, then upload to an S3 bucket.
This script is best for postgres clusters running in EC2 as it makes extensive use of EC2 instance tagging for all configuration options. 
The variables in init() can also be defined manually or via environment variables if preferred.